requirements.txt
asgiref==3.8.1
chardet==5.2.0
diff-match-patch==20241021
Django==5.2
django-ckeditor==6.7.2
django-import-export==4.3.7
django-js-asset==3.1.2
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
reportlab==4.4.0
sqlparse==0.5.3
tablib==3.8.0
whitenoise==6.9.0
